:white_check_mark: ensure RemoveBookUseCase returns error if no book is found

+++ tests/application/useCases/remove_book_test.go
@@ -2,6 +2,7 @@ package usecases_test
 
 import (
 	"errors"
+	"log"
 	"testing"
 
 	"github.com/KPMGE/go-users-clean-api/src/domain/entities"
@@ -22,8 +23,9 @@ type RemoveBookRepositorySpy struct {
 }
 
 type FindBookRepositorySpy struct {
-	FindInput string
-	FindError error
+	FindInput  string
+	FindOutput *entities.Book
+	FindError  error
 }
 
 func (repo *RemoveBookRepositorySpy) Remove(bookId string) error {
@@ -37,7 +39,7 @@ func NewRemoveBookRepositorySpy() *RemoveBookRepositorySpy {
 
 func (repo *FindBookRepositorySpy) Find(bookId string) (*entities.Book, error) {
 	repo.FindInput = bookId
-	return nil, repo.FindError
+	return repo.FindOutput, repo.FindError
 }
 
 func NewFindBookRepositorySpy() *FindBookRepositorySpy {
@@ -57,7 +59,11 @@ func NewRemoveBookUseCase(removeBookRepo RemoveBookRepository, findBookRepo Find
 }
 
 func (useCase *RemoveBookUseCase) Remove(bookId string) error {
-	useCase.findBookRepo.Find(bookId)
+	foundBook, _ := useCase.findBookRepo.Find(bookId)
+	if foundBook == nil {
+		return errors.New("book not found!")
+	}
+
 	err := useCase.removeBookRepo.Remove(bookId)
 	if err != nil {
 		return err
@@ -68,8 +74,15 @@ func (useCase *RemoveBookUseCase) Remove(bookId string) error {
 func MakeRemoveBookSut() (*RemoveBookUseCase, *RemoveBookRepositorySpy, *FindBookRepositorySpy) {
 	removeBookRepo := NewRemoveBookRepositorySpy()
 	removeBookRepo.RemoveError = nil
+
 	findBookRepo := NewFindBookRepositorySpy()
 	findBookRepo.FindError = nil
+	fakeBook, err := entities.NewBook("any_title", "any_author", "any_description", 100.2, "any_user_id")
+	if err != nil {
+		log.Fatal(err)
+	}
+	findBookRepo.FindOutput = fakeBook
+
 	sut := NewRemoveBookUseCase(removeBookRepo, findBookRepo)
 	return sut, removeBookRepo, findBookRepo
 }
@@ -98,4 +111,14 @@ func TestRemoveBookUseCase_ShouldCallFindBookRepositoryWithRightBookId(t *testin
 	sut.Remove("any_book_id")
 
 	require.Equal(t, "any_book_id", findBookRepo.FindInput)
+}
+
+func TestRemoveBookUseCase_ShouldReturnErrorIfFindBookReturnsNil(t *testing.T) {
+	sut, _, findBookRepo := MakeRemoveBookSut()
+	findBookRepo.FindOutput = nil
+
+	err := sut.Remove("any_book_id")
+
+	require.Error(t, err)
+	require.Equal(t, "book not found!", err.Error())
 }
